# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ComicClan",
        "description": "An complete API made to serve new upcoming ComicClan social network",
        "contact": {
            "name": "Jean Melo",
            "email": "jf.melo6@gmail.com"
        },
        "version": "0.0.1-oas3"
    },
    "servers": [
        {
            "url": "https://virtserver.swaggerhub.com/ghanizadev/comic-clan/1.0.0"
        }
    ],
    "security": [
        {
            "OAuth2 Authentication": []
        }
    ],
...
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Shows all registered users.",
                "responses": {
                    "200": {
                        "description": "OK - Operation completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Provided client Id and client secret does not match or they were not set in basic authentication.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2 Authentication": ["profile"]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Creates a user by providing email, name and password.",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateAlterUser"
                },
                "responses": {
                    "201": {
                        "description": "Created - User was successfully added to the system.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Request is malformed, missing fields or contains invalid headers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Provided client Id and client secret does not match or they were not set in basic authentication.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type - Request is contains an unsupported media type. Currently supported are \"application/json\", \"application/x-www-form-urlencoded\" and \"form-data\". Check each endpoint documentation to see what content-type you should use.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Basic Authentication": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete an user",
                "description": "Remove your own account from database.",
                "responses": {
                    "204": {
                        "description": "No Content - User was succesfully removed."
                    },
                    "400": {
                        "description": "Bad Request - Request is malformed, missing fields or contains invalid headers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - It is only possible to delete your own account.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get an user",
                "description": "Shows user's information",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User's unique ID",
                        "schema":{
                            "type": "string",
                            "format": "hex"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Operation completed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - \"profile\" is not included in your scope",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2 Authentication": ["profile"]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Alter a user",
                "description": "Change user's data.",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "description": "User's unique ID",
                        "schema":{
                            "type": "string",
                            "format": "hex"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAlterPost"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created - User was successfully changed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Request is malformed, missing fields or contains invalid headers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - It is only possible to change your own data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type - Request is contains an unsupported media type. It must be \"application/json\".",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - You achieved the maximum requests per second.",
                        "headers": {
                            "Retry-After": {
                                "description": "Request timeout for new requests, in seconds.",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "format": "hex"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "xml": {
                    "name": "PaginatedUser"
                }
            },
...
            "CreateAlterPost": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "xml": {
                    "name": "Post"
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "error_description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "number"
                    }
                },
                "xml": {
                    "name": "Error"
                }
            },
...
        },
        "requestBodies": {
            "CreateAlterUser": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            },
                            "xml": {
                                "name": "User"
                            }
                        }
                    }
                },
                "required": true
            }
        },
        "securitySchemes": {
            "Basic Authentication": {
                "type": "http",
                "scheme": "basic"
            },
            "OAuth2 Authentication": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "https://localhost/oauth/token",
                        "refreshUrl": "https://localhost/oauth/token",
                        "scopes": {
                            "feed": "allows reading feed resources",
                            "post": "allows posting to user's feed",
                            "profile": "allows reading, editing and deleting user's profile",
                            "comment": "allows commenting in post resources"
                        }
                    }
                }
            }
        }
    }
}